requirements.txt
allure-pytest==2.13.5
requests
pytest==8.1.1
locust
pyyaml
pytest-xdist==3.5.0
